refactor: split clients and servers utils
