+++ portal-front/app/sitemap.ts
@@ -1,9 +1,7 @@
 import { serverFetchGraphQL } from '@/relay/serverPortalApiFetch';
 import { PUBLIC_CYBERSECURITY_SOLUTIONS_PATH } from '@/utils/path/constant';
-import {
-  fetchAllDocuments,
-  ServiceSlug,
-} from '@/utils/shareable-resources/shareable-resources.utils';
+import { ServiceSlug } from '@/utils/shareable-resources/shareable-resources.types';
+import { fetchAllDocuments } from '@/utils/shareable-resources/shareable-resources.utils';
 import SeoServiceInstancesQuery, {
   seoServiceInstancesQuery,
 } from '@generated/seoServiceInstancesQuery.graphql';
