+++ portal-front/src/utils/shareable-resources/shareable-resources.utils.ts
@@ -1,163 +1,13 @@
 import { serverFetchGraphQL } from '@/relay/serverPortalApiFetch';
-import SeoCsvFeedBySlugQuery from '@generated/seoCsvFeedBySlugQuery.graphql';
-import SeoCsvFeedsByServiceSlugQuery from '@generated/seoCsvFeedsByServiceSlugQuery.graphql';
-import SeoCustomDashboardBySlugQuery from '@generated/seoCustomDashboardBySlugQuery.graphql';
-import SeoCustomDashboardsByServiceSlugQuery from '@generated/seoCustomDashboardsByServiceSlugQuery.graphql';
 
 import { fromGlobalId } from '@/utils/globalId';
-import { csvFeedsItem_fragment$data } from '@generated/csvFeedsItem_fragment.graphql';
-import { customDashboardsItem_fragment$data } from '@generated/customDashboardsItem_fragment.graphql';
-import { ConcreteRequest } from 'relay-runtime';
-
-export type ShareableResource =
-  | customDashboardsItem_fragment$data
-  | csvFeedsItem_fragment$data
-  | SeoCsvFeed
-  | SeoCustomDashboard;
-
-export type SeoResource = SeoCsvFeed | SeoCustomDashboard;
-
-export interface SeoCustomDashboard {
-  description: string;
-  id: string;
-  type: 'custom_dashboard';
-  children_documents: {
-    id: string;
-  }[];
-  created_at: string;
-  updated_at: string;
-  labels: {
-    color: string;
-    id: string;
-    name: string;
-  }[];
-  name: string;
-  slug: string;
-  short_description: string;
-  product_version: string;
-  download_number: number;
-  share_number: number;
-  uploader: {
-    first_name: string;
-    last_name: string;
-    picture: string;
-  };
-  active: boolean;
-  service_instance: {
-    id: string;
-    slug: string;
-  };
-  uploader_organization: {
-    id: string;
-    name: string;
-    personal_space: boolean;
-  };
-}
-
-export interface SeoCsvFeed {
-  description: string;
-  id: string;
-  type: 'csv_feed';
-  children_documents: {
-    id: string;
-  }[];
-  created_at: string;
-  updated_at: string;
-  labels: {
-    color: string;
-    id: string;
-    name: string;
-  }[];
-  name: string;
-  slug: string;
-  short_description: string;
-  download_number: number;
-  share_number: number;
-  uploader: {
-    first_name: string;
-    last_name: string;
-    picture: string;
-  };
-  active: boolean;
-  service_instance: {
-    id: string;
-    slug: string;
-  };
-  uploader_organization: {
-    id: string;
-    name: string;
-    personal_space: boolean;
-  };
-}
-
-type QueryMapEntry<TReturn> = {
-  query: ConcreteRequest;
-  cast: (data: unknown) => TReturn;
-};
-
-type MakeQueryMapParams = {
-  query: ConcreteRequest;
-  key: string;
-};
-
-type ServiceInfo = { link: string; description: string };
-export enum ServiceSlug {
-  OPEN_CTI_INTEGRATION_FEEDS = 'open-cti-integration-feeds',
-  OPEN_CTI_CUSTOM_DASHBOARDS = 'open-cti-custom-dashboards',
-}
-
-function makeQueryMapEntry<TReturn>({
-  query,
-  key,
-}: Omit<MakeQueryMapParams, 'isList'>): QueryMapEntry<TReturn[]> {
-  const cast = (data: unknown): TReturn[] => {
-    const safeData = data as Record<string, unknown>;
-    return safeData[key] as TReturn[];
-  };
-  return { query, cast };
-}
-
-function makeSingleQueryMapEntry<TReturn>({
-  query,
-  key,
-}: Omit<MakeQueryMapParams, 'isList'>): QueryMapEntry<TReturn> {
-  const cast = (data: unknown): TReturn => {
-    const safeData = data as Record<string, unknown>;
-    return safeData[key] as TReturn;
-  };
-  return { query, cast };
-}
-
-export const localeMap: Record<ServiceSlug, string> = {
-  [ServiceSlug.OPEN_CTI_INTEGRATION_FEEDS]: 'CsvFeed',
-  [ServiceSlug.OPEN_CTI_CUSTOM_DASHBOARDS]: 'CustomDashboards',
-};
-
-const queryMap: Record<ServiceSlug, QueryMapEntry<SeoResource[]>> = {
-  [ServiceSlug.OPEN_CTI_INTEGRATION_FEEDS]: makeQueryMapEntry<SeoCsvFeed>({
-    query: SeoCsvFeedsByServiceSlugQuery,
-    key: 'seoCsvFeedsByServiceSlug',
-  }),
-  [ServiceSlug.OPEN_CTI_CUSTOM_DASHBOARDS]:
-    makeQueryMapEntry<SeoCustomDashboard>({
-      query: SeoCustomDashboardsByServiceSlugQuery,
-      key: 'seoCustomDashboardsByServiceSlug',
-    }),
-};
-
-const querySlugMap: Record<ServiceSlug, QueryMapEntry<SeoResource>> = {
-  [ServiceSlug.OPEN_CTI_INTEGRATION_FEEDS]: makeSingleQueryMapEntry<SeoCsvFeed>(
-    {
-      query: SeoCsvFeedBySlugQuery,
-      key: 'seoCsvFeedBySlug',
-    }
-  ),
-  [ServiceSlug.OPEN_CTI_CUSTOM_DASHBOARDS]:
-    makeSingleQueryMapEntry<SeoCustomDashboard>({
-      query: SeoCustomDashboardBySlugQuery,
-      key: 'seoCustomDashboardBySlug',
-    }),
-};
+import { queryMap, querySlugMap } from './shareable-resources.consts';
+import {
+  SeoResource,
+  ServiceInfo,
+  ServiceSlug,
+  ShareableResource,
+} from './shareable-resources.types';
 
 export async function fetchAllDocuments(
   serviceSlug: ServiceSlug
